[flink][refactor] Move Kafka related code into paimon-flink-cdc module (#1988)
