+++ paimon-docs/src/main/java/org/apache/paimon/docs/configuration/ConfigOptionsDocGenerator.java
@@ -77,7 +77,7 @@ public class ConfigOptionsDocGenerator {
                 new OptionsClassLocation(
                         "paimon-flink/paimon-flink-common", "org.apache.paimon.flink"),
                 new OptionsClassLocation(
-                        "paimon-flink/paimon-flink-common", "org.apache.paimon.flink.kafka"),
+                        "paimon-flink/paimon-flink-cdc", "org.apache.paimon.flink.kafka"),
                 new OptionsClassLocation(
                         "paimon-hive/paimon-hive-catalog", "org.apache.paimon.hive"),
                 new OptionsClassLocation(
