+++ paimon-flink/paimon-flink-common/src/main/resources/META-INF/services/org.apache.paimon.factories.Factory
@@ -13,8 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.paimon.flink.kafka.KafkaLogStoreFactory
-
 ### action factories
 org.apache.paimon.flink.action.CompactActionFactory
 org.apache.paimon.flink.action.CompactDatabaseActionFactory
